fix: scope subagent state and navigation to source threads
