+++ desktop/src/renderer/tests/notificationDispatch.test.ts
@@ -280,7 +280,6 @@ function dispatch(payload: { method: string; params: unknown }): void {
         if (
           entries.length > 0
           && knownChildCount < entries.length
-          && !nextSubAgentStore.loadingParents.has(threadId)
         ) {
           void nextSubAgentStore.fetchChildren(threadId)
         }
@@ -549,7 +548,7 @@ describe('notification dispatch payload format', () => {
     await vi.waitFor(() => {
       expect(sendRequest).toHaveBeenCalledWith('subagent/children/list', {
         parentThreadId: 'thread-1',
-        includeClosed: false,
+        includeClosed: true,
         includeThreads: true
       })
       expect(useSubAgentStore.getState().childrenByParent.get('thread-1')?.[0]).toEqual(
@@ -847,7 +846,7 @@ describe('notification dispatch payload format', () => {
     await vi.waitFor(() => {
       expect(sendRequest).toHaveBeenCalledWith('subagent/children/list', {
         parentThreadId: 'thread-1',
-        includeClosed: false,
+        includeClosed: true,
         includeThreads: true
       })
       expect(useThreadStore.getState().threadList.some((thread) => thread.id === 'child-graph')).toBe(true)
